backend/ocr_donut/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-multipart
pydantic
Pillow
aiofiles
